The feature to block applications with opened popups
If the application has an opened popup, none of its areas may be
accessed by the user. Therefore, Luwrain has a special technique to
block these areas in one of the following ways: totally reject any
interaction or silently close a popup if the user tries to deal with
the areas (so called weak popups, possible only if the popup is a
top-level popup). Meantime, Area.getName() may be called anyway even
if there are opened popups and ThreadSyncEvent is delivered
immediately regardless whether there are opened popups or not.