Brief introduction in open file popup
Pressing the Control key, user is now able to hear only a file name,
instead of listening a full path.  This requires a deep change in
EditListPopup, because the methods in its model to get next and prev
items now return an item class with additional data, not a string.