New approach to util.MlTagStrip
Now MlTagStrip class considers a closing tag only if it has the same
name as opened. Previously it didn't take care of the name of the
closing tag implying that handling code is able to recognize an
error. Now occurring a closing tag with an invalid name is completely
valid situation.

Minor fix:

core.FileTypes gets support of *.doc and *.html