New rules of loading speech channels (--speech= introduced)
LUWRAIN now recognizes --speech= command line option, which must have
a colon-delimited list of string arguments. First of them specifies
the type of the speech channel to create, all remaining are provided
to the channel for the initialization. The speech channel decides on
its own how to handle them.  Using of --speech= blocks loading of
speech channels from the registry, and the speech channel created with
it is always selected as main speech channel for speech output.