Introducing new control panel organization
Now extensions should provide so called factories to declare what
sections they want to add to the control panel. Each factory provides
a list of elements, which can be compared and which reflects the tree
of sections. After that, control panel application may ask which
section to create to be added to the tree. This is needed for
effective tree refreshing operation.

This approach is now introduced, but still incomplete. Control panel
application gets Base and TreeItem classes. All standard sections go
to org.luwrain.settings package. Standard elements are declared
in.StandardElements class.

TreeArea.Model no longer has isLeaf() method. If getChildCound()
returns zero, it means that the node is a leaf.