Class path patching in runtime
core.ClassPath class added for runtime patching of classpath list. We
forced to do this way, because there are a lot of problems with
gathering of jars. It's especially difficult to deal with package.cfg,
which is used with javafxpackager.

Now we are gathering all jars from ./jar and ./lib.